requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.28.0
google-cloud==0.34.0
google-cloud-secret-manager==2.20.1